Information on running HA in background.
I still struggle with getting HA to run as a Daemon but did find the very useful nohup command which at least allowed me to launch HA as a background task.  This could be useful for others until there is a more complete set of instructions for running hass as a deamon or Autolaunch.