Update sensor.sabnzbd.markdown
Added clarification about the name of the sensors created to help new users who might expect the sensor name and component name to match.

Also, it would seem that after 0.27, configuration of this component may have changed.  Using "http" or "https" as indicated in the documentation seems to raise an exception.  Using the port number on the host line also causes an error.  Omitting this (at least without SSL) seems to fix the issue.  I don't know enough about how this component work to troubleshoot, so I left that part as-is.  Maybe someone who knows this component better could take a look?