Update z-wave.markdown
I added this automation when first setting up Home Assistant and the soft reset has been very problematic for me, it hangs my zstick gen5.  So I wanted to make sure that people only add this automation if they run into issues and not add it as part of setting up hass.