Update installation-raspberry-pi-image.markdown
Add info to customize. I had to search a few minutes to figure out how to change the keyboard layout. 

Add Windows 10 SSH in bash info. Maybe there should be a seperate page in the documentation, where all 'How to connect/SSH' topics link to. That way it won't have to be repeated throughout the docs.