Security Tip for changing AiO MQTT password.
went back and forth on whether to put this here or in the All in One installer docs.  I think it is best suited here since users will come here once they want to learn more about MQTT and their particular installation which could be a significantly time after when they first did the install.