Add known issues for the pilight.send service
As we don't implement the pilight api but only provide a transport to
the daemon, we should inform the user, that its their responsibility to
choose the correct data type.

As we convert all the ussers data to string if it specifies it in a
data_template without giving them an option to change it back, we
should put a warning, that templating will not work at all with the
send service of pilight.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>