Clarify docs for binary_sensor class 'opening'
The documentation for the 'opening' class of binary_sensor left the
interpretation of On and Off undefined.  Additionally, the intro paragraph
claimed that "On" is "Closed" and "Off" is "Open" which is the opposite of
common usage that I've seen in example configurations.

This commit explicitly states that On==Open and Off=Closed and amends the
examples in the leading paragraph.

Confirmed in Gitter chat that this is appropriate:

@balloob: "I would assume that “on” is open as it’s the special case.
           Assuming doors/windows etc are usually closed.
           With smoke, detecting smoke is the special case"