Add /bin to PATH variable for virtualenv installations
Add /bin to PATH variable for virtualenv installations to enable home assistant to find e.g. the ping command which is located in /bin. This was a problem on ubuntu which caused the WOL switch to not report the correct state of the computer.

See https://github.com/home-assistant/home-assistant/issues/5930 and https://community.home-assistant.io/t/wake-on-lan-switch-status-not-working/11700 for reference.