add troubleshooting steps to docs
I spent a day wondering why `ubus` wouldn't report changes to the device list in the ARP table before digging into the packets and discovering that it only looks at `hostapd.*`. Since I don't use the AP on my router, this won't work for me. I would have been thrilled to know in advance that it doesn't apply to my use case. I've added a line at the top that clarifies what it looks for, and I've added a block of troubleshooting steps because they were fresh in my mind and might help others in the future. I see a lot of posts in the forums about, "what the heck is <insert component> doing?" Knowing how to look at packets with tcpdump and wireshark is a useful skill.