Proposed update to yaml config
I have not read the docs extensively but the yaml config example given seems incorrect.  It have adjusted to to what I believe is the correct config for the automation example given.  I am not sure if things changed with the new Automation editor but I don't want people struggling with configs due to bad documentation.