erorrs with degree symbol (#2973)
I was unable to get the examples above working when using that degree symbol in the unit of measurement, removing that symbol fixed the problem. I did try pasting in a fresh degree symbol, and even tried using one from the HASS front end, but that didn't work. However, this example has been here for a long time and I can't find any other cases of complaints, so I figured i'd add a warning in the comments in case anyone else has the same problem.