Updated filename for iOS, added token description
The filename to extract the token from the iOS backup has the USERID in the front ({USERID}_mihome.sqlite). A file without the {USERID} existed as well, so i think the clarification is needed.
In addition i added a description of how the token looks so one knows what to look for.