Added note about updates for Hass.io users.
Hass.io users end up on this page when an update is available, but then can't see what they have to do.  This is because of the the updater component and the inevitable delay in updates getting into the hass.io build.

This update informs users about an updates transition from HA into Hass.io, how to install it, and suggests disabling the updater component to alleviate 'notification stress'.