Add documentation for availability topic and payloads (#3378)
* Add documentation for availability topic and payloads

Added documentation for availabiltiy topic and payloads added in a pull request. Also cleaned up documentation for the MQTT cover generally to make clear the differences between state topic and messages and availability topic and messages.

* Add default qos level

* Add documentation for availability topic and payloads

Added documentation for availability topic and payloads added to the MQTT binary sensor in a pull request. Also cleaned up documentation for the MQTT binary sensor generally.

* Remove sending