Update switch.rpi_gpio.markdown (#3662)
I found the current documentation confusing as it did not clearly state whether the code is look at the pin # or GPIO #. Looking on the forums others were having the same issue. My goal is to make it clear that GPIO # is what users should be entering into their configs.