Update python_script.markdown
Using Homeassistant 0.56.2, the latest at this time, the component python_script.py defines the DOMAIN as 'python_script' and the FOLDER 'python_scripts'.  As is you will encounter the error: 'Unable to find component python_script'.  It took awhile to figure out that figure out that the DOMAIN and FOLDER should be linked and a quick change to the documentation was all that was needed.