Change link to tutorial to get the key
The old links were referring to a forum post in which it only was explained how to enable the advanced settings, followed by a link to the remaining part of the tutorial (on the Domoticz site). I changed the link to the domoticz forum directly, as it gives the full instructions, including enabling the advanced settings in the app. Added the word “password”, as it is called “password” in the Mii app, and not “key” in the app.