Clarify advanced install instructions
I just added a direct command to run for an advanced install. The docs at https://github.com/home-assistant/hassio-build/tree/master/install#install-hassio need some cleanup too, but that is clearly a separate change. 

I'm not a huge fan of adding code here that might change, but otherwise the linked document is confusing out of context, and I suspect that set of commands isn't actually that likely to change.