Display problem with script section...
I banged my head against the table for hours trying to make the script work on this page and then I figured out something.  When viewing the page on the HA site, the variable tags {{value1}}, etc...do not display.  I only see two quotes.

This is what I was seeing.  ```{"event":"TestHA_Trigger", "value1":"", "value2":"", "value3":""}```

And I should be seeing this:  ```{"event":"TestHA_Trigger", "value1":"{{value1}}", "value2":"{{value2}}", "value3":"{{value3}}"}

Above this section in the automation, there's a tag called { % raw % } which I'm unfamiliar with, but I made a guess that this needs to be inserted.  I may have been wrong.  If so, I hope I've identified a valid problem.  Thanks!