Update sensor.template to show entity_id example (#4158)
Right now this page has no examples using the optional entity_id field. I think we should include it in the top example because otherwise the template sensor will react to all state changes rather than just the one or two the user will care about. May also be nice to show an example of a list for the template sensor, but we do include one on the Binary Sensor Template page so maybe that's sufficient?