Update Entity Namespace Information
There are still core platforms within Home Assistant that do not fully support the entity_namespace attribute. This update will establish the requirements for supporting this attribute (entity_component class), as well as a list of core platforms which do and do not support this attribute. device_tracker is the primary culprit that has been identified at this time. A feature request will be submitted for this, and this page will be updated if the change is made to the component.