Fix some leftover references to the switch component
The documentation for the mochad light component was mostly copy and
pasted from the mochad switch device. Since the configuration is more or
less the same between devices. However in doing this several references
to mochad switches was left in by mistake. This commit fixes these and
replaces them with the correct name of the component.