Added `switch:` to configuration.yaml example
Added `switch:` to configuration.yaml example to help clarify what actually needs to be added to a configuration.yaml file in order for this platform to work- assuming this is the only switch in the configuration, or if you are adding it as a list item to other switches. It cannot be copied/pasted directly as it was.