Clarification regarding the frequency...
It's often been noted in the forums that Template Triggers are sometimes overload and that you want to use them sparingly.  I haven't stress tested my system with this idea, but if the template triggers are tested every time any state changes, then that is the clarification that I've proposed in the documentation.