Update binary threshold sensor configuration
The existing documentation is incorrect and results in config errors. Per this thread (https://community.home-assistant.io/t/threshold-binary-sensor/7316/5) and my experience, my file change is in line with what the configuration actually needs to be.

My proposed changes are probably incomplete due to the fact that I'm not a programmer, but I hope it's more useful than the erroneous info in the existing version!