Adding configuration tag according to new docs
One thing is missing (need help!): I did not find a sample of how to specify the keys for the map type in configuration. The component accepts a list of states that can contain lists of codes, like this:
report_server_codes:
        arm: XXXXXXXXXXXXXXXX, XXXXXXXXXXXXXXXX
        disarm: XXXXXXXXXXXXXXXX, XXXXXXXXXXXXXXXX
        armhome: XXXXXXXXXXXXXXXX
        triggered: XXXXXXXXXXXXXXXX, XXXXXXXXXXXXXXXX, XXXXXXXXXXXXXXXX
        ignore: XXXXXXXXXXXXXXXX

not sure how to put this in the %configuration% tag. For now, I left both the old version and the {%configuration%} tag in.