Pushover: update webpage with clone URL
Hi, Pushover developer here.

I was notified that your app has a Pushover plugins, so I've listed it on our [apps page](https://pushover.net/apps) and enabled cloning for it.  Instead of users having to upload an icon and fill in all the fields, users can be directed to https://pushover.net/apps/clone/home_assistant which pre-fills those fields and copies the icon.

I'm not sure if the `date` field in this markdown page needs to be bumped when it's updated, or if it's a creation date.