specify security concerns and actions to mitigate them.
With the focus on security concerns in the forums recently I think it would benefit users to be more aggressive in recommending actions to take to secure home assistant when exposing it to the internet. Vague warnings don't always catch the eye of the casual user. I was one of those users and the forum threads on security caused me to examine my own security and I then realized how potentially lacking that it was.