Update device_tracker.ubus.markdown
Several HA users encountered a problem after upgrading OpenWRT to version 18.06.x. UBUS did not function anymore (TypeError: ‘NoneType’ object is not subscriptable). At https://community.home-assistant.io/t/problem-with-ubus-device-tracker/46072/18     user rf3141 analyzed the problem (besides the https issue the thread was started for) and solved it by installing the additional package uhttpd-mod-ubus. Maybe this needs to be analyzed more careful but for now the addional package solves the problem.