Update tuya.markdown
Values should be enclosed in single quotes so I updated the example accordingly. Platform required was set to false; it is required so I changed it to true and added it to the example. Only two platform options were given- tuya and smart_life. My devices required the platform to be set to 'jinvoo_smart' so I updated the platforms to include that option. If there are other known options, I suggest they be listed as well.