Port using for MariaDB 10
The text says that when using a MariaDB 10, you also need to add port 3307 in the link. If you use an external database that is, for example, run at a hosting provider, you have to be careful that you use the correct port! I had assumed (because my database is also MardiaDB 10) that I had to use port 3307, but on inquiry it turned out to be 3306. Hence my adjustment.