sudo to homeassistant before virtualenv is activated on Fedora
Very similar approach to Debian-based installation needs to be done
here. In the previous block of commands user homeassistant is created
and directories are chowned to him.

If one would try to run followup pip3 install or even hass itself as
current user he wouldn't be able to write to those directories or that
homeassistant user would be just useless.