Clarification about the region name
Myself and, by searching for a solution to my wrong usage of `region_name`, others have misunderstood - or not read carefully - the documentation. We used the name of our city by prior searching for it on the DWD website. This mistake is definetly on our part but could hopefully avoided by stronger pointing out the existing section in the documentation.

I surrounded the section by an `important` box and rephrased the wording a little.