Updated to reflect the new API limits
The free tier of OpenUV is now limited to 50, not 500.  All existing accounts will drop to that limit in February, but all new accounts have that limit.  The current scan_interval of 2/hr puts it at 48 calls/day (without any reboots triggering extras).  Should we adjust the default to be 1/hr or give documentation on how to not query it until the sun is above the horizon?