Add webhook example to trigger.markdown
This makes more clear that: 1) https may be needed rather than http, 2) data is not needed for a webhook trigger to function, and 3) "curl" is a good way to test it.  Hoping to save other people some of the time I spent figuring out how to get it to work.