Removed extra key "authentication"
Received this extra key warning:

Your configuration contains extra keys that the platform does not support.
Please remove [auhentication].  (See /config/cameras.yaml, line 0).

The authentication key was still in the documentation.

Removing the key from my configuration did not affect correct operation of the camera. Assume this is now an invalid key so have submitted this PR to remove it.