Add troubleshooting section
When I first attempted to setup the stream component in my dev environment I had the following errors that were only solved by installing these packages.  When I moved to a second pi to do further testing I again needed to run the same commands to setup the component so this will be helpful to any user who has the same error.