Fix formatting of installation commands
This change fixes the formatting of the generic linux host
installation commands so that they all begin with `$`.
This ensures that they are rendered correctly in the documentation.

Previously, the commands under `$ sudo -i` all began with `#` (to
indicate that the shell was running as root). While this is correct,
the markdown formatting thinks that they are comments, so they
are rendered in a light gray color.