Describe support of multiple elk devices (#9441)
* Describe support of multiple elk devices

For an upcoming PR that supports multiple elk devices.  This is a doc update for the version of that PR that uses multiple host sections instead of a list of host sections under a devices tag (since that extra "devices" section would result in a breaking change).

* :pencil2: Tweak

* Added a prefix field to one of the config entries

The example config needed a prefix for the second entry.