Add De Lijn component documentation (#9507)
* Add De Lijn component

Documentation of the De Lijn component to add passages of stops of the Flemish public transport company into home assistant

* :pencil2: Tweak

* Update source/_components/delijn.markdown

Suggested corrections applied

Co-Authored-By: Klaas Schoute <klaas_schoute@hotmail.com>

* Some suggested changes added

Sensor added as ha_category.
Only show the mandatory items in the configuration example.
Rewritten the Configuration part a little bit to align with example.
Added some additional information on the Full Configuration to clarify the outcome of that configuration.

* Alignment of configuration names

Aligned the names in the configuration explanation with the actual configuration names.

* Linked to custom lovelace card added

Added link to custom lovelace card that makes best use of all the retrieved features (ie. correct colour codes of the line cards)

* Apply suggestions from code review

Co-Authored-By: Klaas Schoute <klaas_schoute@hotmail.com>
Co-Authored-By: Franck Nijhof <frenck@frenck.nl>

* Documents aligned with changes in component code

sub_key was replaces with api_key in the component code so the reference is changed in the documentation here as well.

* Updated with changes to component

The delijn component has been updated based on feedback given in PR #24265. Based on those changes the documentation had to be updated as well. Now documentation and component are aligned again.

* :pencil2: Markdown cleanup

* :art: Adds De Lijn logo