correct the quotes in continue_on_timeout: 'false'
docs show continue_on_timeout: 'false' while this doesn't stop the script from continuing. using continue_on_timeout: false does work correctly, so the quotes should be taken out on the false.
haven't tested this on the continue_on_timeout: 'true', but figure that to be the same. (I won't PR that just yet, just to be 100% sure)

posted on this in the community on https://community.home-assistant.io/t/script-help-wait-template-and-proceed-after-x-minute/56158/12?u=mariusthvdb