Document 'name' config parameter in rejseplanen (#10243)
* Document 'name' config parameter in rejseplanen

Document the undocumented but fully working name config. This is an optional config but essential to know to define more than one instance of the rejseplanen sensor and getting a predictable result. Otherwise the sensor name defaults to sensor.next_departure_# and the numbers can swap around after an HA restart

* An example where policy and common sense clash

Done as suggested though I do not agree

* :pencil2: Tweak

* Adjust default

* Fix sample