Update counter.markdown
I landed on [this page](https://community.home-assistant.io/t/counter-warning-error-issue/85490) in search for a solution to the same problem. The solution was adding "system_log:" and "fire_event: true" and this is the proposed change to the article.

The example configuration I added in case it is helpful to others. 
Have a great day, you, whoever is reading this and either rejecting or approving this PR :)