Clarity on emabling MQTT TLS (#10453)
Clarified that you MUST define the `certificate` parameter to enable TLS connections to the MQTT server. This was not clear to me and was fighting with getting the right PEM/CRT file setup as commections were failing. When I then selected 'auto' it worked as designed. It would be better if there was an 'enable_tls' config parameter that would enable it and default to 'auto', but that;s a code change.