Update zha.markdown with more detailed use-case description
Update zha.markdown with more detailed use-case description about its HA native functionality.

This fleshing out of the description is to try to better explain to end-users (as well as to developers) that from an end-users point-of-view this Zigpy integration adds a native Zigbee function implementation to HA, (similar to what the OpenZWave integration does for Z-Wave in HA). I think that this or a similar fleshed out description is needed so that end-users that are new to HA will know the difference between this integration as others that use third-party hardware or software application acting as a gateway or hub, such as example DeConz, Zigbee2mqtt, Philips Hue Hub, and Ikea Tradfri Gateway.

Also added a link to ZiGate website and clarified its firmware version requirement.