Add link for zigbee.markdown (starter Zigbee doc) to index.markdown
Adding a link for zigbee.markdown (starter Zigbee doc) to index.markdown. Depends on #10582 being accepted first.

Hope that this will help Zigbee become more of a first-class citizen in Home Assistant, just like Z-Wave already is.

The purpose is to expose the Zigbee Home Assistant integration and show in same light as the Z-Wave integration.