Add link for zigbee.markdown (starter Zigbee doc) to index.markdown (current branch)
Adding a link for zigbee.markdown (starter Zigbee doc) to index.markdown. Depends on #10585 being accepted first.

The purpose is to expose the Zigbee Home Assistant integration and show in the same light as the Z-Wave integration. I hope that this will help the existing Zigbee integration become more of a first-class citizen in Home Assistant, just like the existing Z-Wave integration already is. 

One of the reasons for me creating a separate PR for this from #10585 is to give you the option of accepting that and rejecting this index PR if you so wish. As a non-developer, I also currently am not good enough with Git to edit two separate files (using GitHub web interface to edit and create files) :P

PS: This PR is for the 'current' branch and replaces #10583 which was based on the 'next' branch.