Add link for zigbee.markdown (starter Zigbee doc) to index.markdown
Adding a link for zigbee.markdown (starter Zigbee doc) to index.markdown (current branch). 

The purpose is to expose the Zigbee Home Assistant integration and show in the same light as the Z-Wave integration. I hope that this will help the existing Zigbee integration become more of a first-class citizen in Home Assistant, just like the existing Z-Wave integration already is.