Update glossary.yml (#10629)
It came up [in this thread](https://community.home-assistant.io/t/is-there-any-basic-tutorial-on-hassio/140714/5) that the docs don't explain things like what a binary sensor is in an easy to find way. I've pulled a bunch of common platforms into this, not all, and am open to suggestions as to what others I could/should add.