Update link to speedtestservers (#10612)
* Update link to speedtestservers and server_id 

The link to the speedtestservers seems to be broken. The new link does work. Also the server_id seems to be mandatory at this point. So this sets the server_id to be required and in the example configuration.yaml inserts an example server_id.

* Removes sever_id from configuration.yaml

Removes sever_id from configuration.yaml. It's not necessary if server_id is optional.

* Changes server_id to be optional again