Docs are contradictory
A couple of minor changes toward improving config clarity.
It says to set username and password before setting site but setting the `username` parameter here in yaml, shows up as an 'invalid parameter' in logs.  Looking at the code, username and password do seem to be legit parameters but they're not listed as valid params (in yaml example above), so not sure what going on for manual yaml config.
Creating a new integration through UI, does ask for username and password.
 
Further docs modification necessary.