Correct grammar and improve clarity
The current version of this documentation refers to the Raspberry Pi 3 B+ as "the latest version" which is no longer correct with the release of the Pi 4. I changed this to explicitly refer to the 3 B+ instead of saying "the latest version." Do we want to continue saying "the latest version" and change the other references to the Pi 4? The Hass.io installer for the Pi 4 is still an RC which is why I ask.