Add support for more Orange Pi Devices
Use only one 'pin_mode' per device instead of groups. This way it will be future proof if the library changes something for just one device instead of a group.

This is backwards compatible as the 'Lite', 'One' and 'PC' are internally handled the same way. 'Prime' is now handled slightly different due to changes in the library, so users of the Orange Pi Prime are advised to change their configuration to this new value.

Also fixes a link anchor.