Update 'other machine types' section in docs (#11228)
I noticed the list of machine types supported by `hassio_install.sh` wasn't up-to-date with the list on [Github](https://github.com/home-assistant/hassio-installer). I updated the list and added a reference to the Github page. Maybe the duplicate list should be removed from here and instead add a reference to the Github?