Update wording and provide new screenshots (#11180)
* Create trash

* Add files via upload

* Delete trash

* Update wording and screenshots

* Update screenshots and wording

Also improve clarity

* Revert improper references to Lovelace.

* "states screen" > "Home Assistant web interface"

* Delete automation-new-action.png

* Delete automation-new-blank.png

* Delete automation-new-name-trigger.png

* Delete integrations.png

* Create trash

* Add new screenshots

* Delete 01 - username.png

* Delete 02 - location.png

* Delete 03 - location.png

* Delete 04 - devices.png

* Delete 05 - lovelace.png

* Delete 06 - config.png

* Delete 07 - integrations.png

* Delete 08 - new integration.png

* Delete 09 - config 2.png

* Delete 10 - automation editor.png

* Delete 11 - new automation.png

* Delete 12 - new trigger.png

* Delete 13 - action.png

* Delete 14 - server control.png

* Delete trash

* Update screenshot paths

* Update onboarding.markdown

* Correct image paths

* Update automation.markdown

* Reword and provide additional information

This information was outdated, especially with the addition of default_config. Hopefully these changes make it clearer for new users who are confused about the missing homeassistant: line in their default configuration (this is something that has been popping up in the forum lately).

I linked to an external source for the time zone names. Perhaps we should add this list to Home Assistant website instead (or if it already exists, my apologies: I can point it to that instead).

* Update configuration.markdown