Calendar offset field usage update
Currently the google calendar integration requires modifying
event titles in order to trigger an event at an offset to the
actual start time.

Its not always possible to adjust event titles (eg shared/business
calendar).

This commit updates the documentation to reflect a new usage of
the offset field which allows specifying the offset directly
in the configuration and modifying calendar events.