add extract instruction to getting started guide
It is not obvious that the downloaded file needs to be extracted before flashing it. I assumed balenaEtcher did it automatically and spent an entire afternoon trying to figure out why the raspberry pi was not booting up. Hopefully, this change will prevent someone from making the same mistake in the future.