Clarify which name to put in devices (#11715)
* Clarify which name to but in devices

When setting up Automatic, I found it confusing which name to put in the `devices` section, and put the nickname in, which broke the integration. Added a note to clarify for future users.

* Resolve @frenck suggestions.

Yeah, I figured the tone was a bit too playful. Figured I'd give it a go though.

Co-Authored-By: Franck Nijhof <frenck@frenck.nl>

* Resolve @frenck suggestions.

Co-Authored-By: Franck Nijhof <frenck@frenck.nl>

Co-authored-by: Franck Nijhof <frenck@frenck.nl>