Condense configuration for easier reading
This is what I use in lots of my custom Lovelace cards and I find it much easier to read through than what we are currently doing and wanted some feedback on what we can do differently. Right now, the Lovelace configurations are a mess and nearly impossible to read with lots of redundancy even on the same page at times with things like the actions. Feedback/Ideas very much welcome.