Change references to Hass.io to Home Assistant
I made a pass at rewording the Hass.io installation page to use the new branding of Home Assistant instead of Hass.io. I even included a brief explanation at the top of the page introducing the branding change, which includes a link to the blog post explaining the change.

I suspect you may not want to make the naming changes one page at a time. I'm wondering if a separate branch of the site needs to be worked on so the changes can be made all at once (including reorganizing the paths to exclude references to hassio).