Update zha.markdown with intro to the ZHA integration and Zigbee in general
Update current zha.markdown with more introductory information which describes the purpose of this ZHA integration/component, the Zigbee concept in general, basic Zigbee troubleshooting, as well as some links ((with short descriptions) to a few related projects and other directly relative integrations. 

I also included a couple of general benefits arguments to compare the ZHA integration with other solutions for getting Zigbee devices into Home Assistant, but without in any way trying to put down the other possible solutions for gettings Zigbee devices into Home Assistant.

In my opinion, all this text in this pull request gives a better introduction to the ZHA integration for those who are either new to Zigbee or those who know about Zigbee but are new to Home Assistant and might be looking as other solutions to replace their current way of controlling Zigbee devices.   

Credit to much of this information goes to this blog post by Tinkerer as well as the contributors to the Zigbee2mqtt.io website who I think have done a great job at describing the benefits of different Zigbee implementations that may not be clear to users without this information if you are either new to Zigbee or new to Home Assistant. 

For reference see:
- https://blog.ceard.tech/2019/11/zigbee-and-home-assistant.html
- https://www.zigbee2mqtt.io/
- https://github.com/ioBroker/ioBroker.zigbee/blob/master/README.md