Update parameter name in Modbus climate (#12083)
This PR updates the documentation to reflect the actual
code for Modbus climate.

An optional parameter, count, is actually named data_count. After discussion whether we should fix it in the HA Core or in the docs, it was decided to fix it in the docs for the sake of not breaking the code for existing users.