Added SAN to the example openssl command
This ensures that the Home Assistant Companion on Android 10 (this is what I tested) will correctly connect to the local Home Assistant instance. This is due to compliance with RFC 2818, which states that support for CN fallback is deprecated and only SAN should be used to establish whether a certificate is valid or not.