Update amcrest doc for new "ptz_control" service (#12383)
* Update amcrest doc for new "ptz_control" service

included full description of ptz_control service and its attributes along with a complete picture-elements card example showing functional arrow movement icons. 

Also mentioned this component can be used for Dahua cameras.

* Update source/_integrations/amcrest.markdown

Co-Authored-By: Franck Nijhof <frenck@frenck.nl>

* Update source/_integrations/amcrest.markdown

Co-Authored-By: Franck Nijhof <frenck@frenck.nl>

* Update source/_integrations/amcrest.markdown

Co-Authored-By: Franck Nijhof <frenck@frenck.nl>

* Update source/_integrations/amcrest.markdown

Co-Authored-By: Franck Nijhof <frenck@frenck.nl>

* Update source/_integrations/amcrest.markdown

Co-Authored-By: Franck Nijhof <frenck@frenck.nl>

* removed "top/bottom" alternate movement commands.

* implement Phil's corrections/suggestions

Also updated the picture-elements card to add a zoom-in/out icon in the center of the arrows.

New example image will also be uploaded once I figure out how to do that in the markdown.

* attempt to create an amcrest folder (for images)

* delete empty amcrest file.

It needed to be a folder.
yet to figure out how to make it so.

* create /amcrest/amcrest_ptz.jpg 

create /amcrest/dummy file in the hopes this creates a new folder

* Upload new amcrest_ptz image

finally figured out how to create a new folder in multiple steps.

* fix ptz_control table formatting

Problem was copying the ptz control markdown from Foscam doc which had 2 columns, not 3.
(all this -|-|- lingo was new to me, and thought copying what worked for one, would work for the other)

Also moved the 2 usage notes to their own subsection

* Delete amcrest_ptz.JPG

deleting because extension is all caps, not lc.    did not see a way to rename a committed file.

* amcrest_ptz.jpg uploaded

* :pencil2: Tweak

Co-authored-by: Franck Nijhof <frenck@frenck.nl>