Wording change to explanation of the actions (#13115)
The wording of the action configuration variables was misleading. The way it reads is that the action will happen once the cover is in that position. I should read as the actions will cause the cover to be in those positions.
Ex. : open_cover "Defines an action to run when the cover is opened."
Most people would read that to mean once the condition is satisfied of the cover being open, the action will then run. It should be that the action will cause the cover to open. I made these changes to the 5 action variables to hopefully make that clearer as it confused me when I was looking to create a cover template.