Yeelight: Minimum brightness incorrect
The minimum brightness when using yeelight transitions is 1 and not 0. This can be seen in the python-yeelight docs (https://yeelight.readthedocs.io/en/stable/yeelight.html#flow-objects) on which this seems to be based. Entering 0 as brightness results in the following error on my machine:
Unable to set scene: {'code': -5000, 'message': 'general error'}