fixed broken(?) manual and made it overall easier
To me the given steps for BlueStacks didn't work. After installing the most recent version of BlueStacks from the given link and installing the most recent version of BS Tweaker from the given link, and installing Mi Home 5.0.30, I was facing a) a setup where BStweaker told me to be incompatible with the given BS Version and b) a not working version of Mi Home. Mi Home was starting but crashing all the time.
Not usable at all.

So I decided to use v5.4.49 like given for non-rooted android. It worked like a charm and I had not to use BStweaker or SQLite Browser at all.

As v5.4.49 didnt work on my mobile this might be a point to merge the two sections or something like that, but as this is my first contribution to home assistant, I dont wanna mess up anything.