Add documentation on for new Influx 2.0 support
Documentation additions to go along with [Influx 2.0 PR](https://github.com/home-assistant/core/pull/35392).  Also small adjustments to the 1.xx documentation to mention the fact that the `where` option in the sensor supports templates and add a new `database` option to the sensor platform to allow you to change the default db for all sensors.