docs(sentry): Add missing Sentry config step
I added the Sentry integration as documented here before this change but it wouldn't actually send events until I noticed the prompt to finish configuring it in the UI.  Once I added the DSN in the UI it started sending events to Sentry.  I confirmed this by adding invalid configuration (`recorder.db_url` with non-existent directory) and no events were sent before adding the DSN in the UI and they were sent after adding it in the UI.

I then tried removing the `dsn` field from `configuration.yaml` figuring it was superfluous.  In fact without it set in `configuration.yaml` the sentry integration failed to start and there was an error message in the UI notifications.  It seems like a bug to require the DSN both in the `configuration.yaml` and the UI.