Added IGMP no response information
I had an issue in the past where my router was blocking IGMP which resulted in an unstable HomeKit connection with Home Assistant. It took me days to found the solution. After I shared the solution on the form I noticed others were helped too. So I think it is useful to put it in the documentation.