Update ubus.markdown
Made more clear that the ACL file should be created on the OpenWRT device and not on the HA. For many experienced users this is evident but for a beginner it's more clear.
After updating Openwrt the ACL file will be deleted. If you do not know this it can take hours to find out why the integration is not working anymore, so I added this to the page.