Fix callback samples
Fix some of the callback button samples:
"Remove this button" button - the space in the command caused the trigger to never fire
"Edit message" button - The value coming back in trigger.event.data.data is "/edit_msg" which, when then trying to send that value back to telegram in the message body, is being treated as markup. It causes the error "Error editing text message: Can't parse entities: can't find end of the entity starting at byte offset […]" and the edit never happens.

NB: I'm not sure if including the replace() is overly confusing. It might be better to just rename the /edit_msg command to /edit to avoid needing to understand what's happening in the template. It could also be explained in the text describing the sample, but it again just feels off-topic.