Fixed appID info and clarified GET request process
Removed statement that appID is specific to an individuals Roku. From what I've been able to find out, the appID remains constant. Technically, you could find the appID by going to the Roku Channel Store and looking at the channel's URL (the appID is contained in it at this time), but sometimes the search won't return results for a channel that exists. Fetching via GET request (i.e. loading the page in your browser) is easier and more reliable a method.