Update full example for modbus binary_sensor
Someone did a bit of a drive by on the doc discord yesterday pointing out that the _full_ example with `scan_interval` hadn't been updated to reflect changes from home-assistant/core#30004. The first example was updated but not this one. I've also tweaked it to more closely match the top example

The user haven't been back since to see they could of made the change, so here it is.