Makes a clearer distinction between this card and mini-graph-card
When configuring from the Lovelace UI it was not immediately clear that the sensor card could not be configured in the same way as the custom mini-graph-card. It is my hope that this note in the doc will save someone else some time in future and possibly point them to a similar, more flexible card.