Adding suggestion to put ',' between hex as well (#14270)
Right now the suggestion is to put '0x' in front of the generated hex keys which is great, but when I followed the instructions I didn't realize it didn't put the comma between the items already and it caused me problems for awhile until I realized. I'm proposing a change to also suggest adding a comma so people, like me, following the instructions will see it.