Added a section for users w/ 2FA on their NAS
The frontend configuration option does not provide a verbose error is you have 2FA turned on for your NAS and haven't yet configured 2FA for the separate user. The only way to know is to dig into the logs where you'll see a reference to OTP but even then it isn't clear what you have to do. Added some context around this to improve the user experience.