Update alexa.smart_home.markdown (#14495)
There look to have been a number of changes to the process so I've tried to document these as best I can. I have not included screenshots but the text I have provided is indicative of the setup experience I've had today (14/09/2020). I also changed some of the wording / sentence structure as it looks like English wasn't the first language for the person who wrote some of the content.