Update index.markdown
When the download is complete for HA, it's not clear that the user isn't immediately taken to the GUI login. Many people with prior experience in Raspbian may be expecting to go straight to the "Onboarding" screen _within their Raspberry pi device_ rather than needing to access it from a separate browser.

Added a suggestion for the last few steps of installation to make this more explicit, but may still benefit from further tweaking / clarity.