Update template.markdown
The template sensor configuration documentation gets completely lost when doing a search in the HA integrations listing.

When you do a search the "Template Sensor" doesn't show up in the list. And it's not obvious when looking at this page that it's actually describing the configuration options for a "template" platform in the "sensor" domain.