Update xiaomi_miio.markdown regarding searching for token
I spend quit some time searching for the correct token. I was not able to do this with described way of working in the alternative methods overview (when having a gateway and an IR remote). I've updated the Android related information but did not delete the 'Bluestacks' solution (which I could not get to work, might be linked to my knowledge (although I have some computer background as programmer for some years)). I might be removed to simplify documentation, any advice on this?