Update Alarm Control Panel With IOT Class
This is my first pull request for Hacktoberfest this year.  I was instructed to mention issue #14661 in this as well.  I tried to look at other source files to match the format and order.  Please let me know if I need to change anything, and I would love to start doing this more:)!