Fixed Broken Links (#14885)
I am not the most proficient with reading python code, but after searching through the Home Assistant Website for a while, I think that I found the Python API article that replaces the broken link.  The server-sent events link is also broken, but there is something weird happening with the link.  I think that I will open an issue for it and then submit a PR with the fix once it is fixed.