Remove obsolete "how it works" page from Lovelace sidebar
This removes the "[How It Works](https://www.home-assistant.io/lovelace/how-it-works/)" link from the Lovelace sidebar navigation. The actual page will still exist, this simply demotes it from the sidebar nav for now.

Reasoning is, the "How it Works" page is obsolete and will only cause confusion and frustration. New users click on "How it works" expecting to see a user guide for Lovelace, but instead they get some nerdy tech document comparing it to the older "States" UI which no longer even exists. 

I would like to eventually rewrite the page to be more user-focused and then re-add it to sidebar.