Create input_schedule.markdown
This is a new integration which should help using fixed automation rules, while controlling their behavior through this new entity type.
The "input_scheudle" entity has an on/off state based on the time periods provided as input from the user. It preserves the time periods during reboots.
A typical usage will be to create a static automation rule, which will behave according to changes in the time periods of the schedule, without the need to change the rule itself.