Rephrase from Home Assistant "Container" to "OS"
The previous phrasing made me (and seemingly many others too) believe
that the Home Assistant Container installation method provides a full
_Supervised_ install of Home Assistant. I think the main issue for me
was the use of the word "image", which I very much associate with
Docker, and containers in general.

I changed the wording to clarify the distinction between the Home
Assistant OS method and the Home Assistant Container method.