Update nest with Device Access program setup
Move the google_nest documentation into the nest integration documentation,
since this is now implemented as part of the existing integration.

The nest documentation is updated to explain the new integration, and
it is listed first (even though it has fewer features) since it is accepting
new users.  The current strategy is to update documentation for the platforms
as is, since over time each platform and most of the configuration options
will be supported by the new API.