Update roomba.markdown
Following the manual of this integration, it is too hard for unexperienced users to get the BLID and Password credentials with the current links on the bottom of the page: https://www.home-assistant.io/integrations/roomba/

I'm not an expert but I'd like to contribute with this solution to make it more easyer to get the BLID and Password. If it's not appriciated, or not according to the rules, then never mind. If this looks good we'll then publish! :-) Thanks and have a great day!
The only change is the extra text on the bottom. The rest is forked as-it-was.