Update Nest config snippet
The current sample yaml for the Nest SDM API asks users to supply subscriber_id as per "Subscription ID in Google Cloud Console".

This is potentially ambiguous - should they provide the Subscription ID, or just replace the last part of the dummy value with it? Users could incorporate "Subscription ID" as show in the console and craft the full path by hand, but they would also need to fill in the missing project ID (represented in the example yaml as '...'). To add to the confusion, the GCP project ID they would need to supply midway through this full path is different to the Device Access Console project_id they have just provided in the previous yaml parameter.

We can simplify this significantly by removing references to "Subscription ID" and just telling users to use the value under "Subscription name" in the same table - this is the fully qualified path which HA really wants. 

Screenshot here for clarity: https://i.imgur.com/on39AtX.png