Update tplink.markdown
I just recently bought a TP-Link KP115 plug to monitor energy for my dishwasher since I had so much success monitoring energy for my washer with a HS110.  Using the same process I used on the washer, I was able to get energy monitoring for my dishwasher with the KP115.  I wanted to let the community know that there is another option for energy monitoring now that the HS110 is discontinued.