update ping device_tracker example and fix variable type
I've clarified the ping integration docs a little bit. I could not figure out
from the `device_tracker` example and the text that `hostname` (under `hosts`)
was meant to be the device name instead of the actual string `hostname`. Also,
the type is wrong, `hosts` is not a list but a map.