Add an example automation
I struggled a bit getting started with automations for this event
because I didn't know the event_type would actually be a `nest_event`.
This made me think that others might want to get a bit of a short cut
working with these events, so I am sharing my automation here to help
the next person along.