Update controllers.markdown
For the Aeotec ZW090 stick, the A, B and C models refer to the radio frequency for US/EU/UK markets and applies to both the old incompatible version, and the new 2020 version that is compatible with the Pi4. Aeotec is calling the new 2020 model is called the "Gen5+". There are sellers on Amazon selling the old Gen5 listing "ZW090-A" as the model number so I feel that "Gen5+" is a more relevant term for a potential buyer to look for for when shopping for this ZWave controller.