Feature Lutron Caséta now that it has a first class config flow
We recently added one touch pairing and support for pico remotes to
Lutron Caséta. Previously this was one of the more difficult
integrations to setup, and was likely quite frustrating for
many users. It would be nice to show we are making strides to
improve the UX and ease of setup.