Update SIMPLISAFE_EVENT results and automation usage.
The original documentation shows SIMPLISAFE_EVENT being an event_type and also serving an event_type as a subset. The current version has an event_type of SIMPLISAFE_EVENT, and the event_data returned is almost all prefaced with "last_". Since this document deals with v2 and v3, it's possible the original documentation displays what a v2 system will show. If that's the case, we should include both sets of event data with a header describing which goes with which Simplisafe version.