Add sensor example for this integration
Actual example of how to use the sensors for this integration are tough to find, so I added one that I figured out.  It would be nice to find a list of valid states for the brother printer integration, but I suspect that is largely determined by which printer you are actually using.