Update systemmonitor.markdown
There are a lot of users who are using the exact yaml from the page. 

They are checking disk space in /home, when in fact they need to check the space in / For many users the numbers are the same, so they don't understand that they are using it slightly incorrect. 

Here is a GitHub search showing some of that incorrect use cases: https://github.com/search?q=systemmonitor+disk_use_percent+arg+%2Fhome&type=code

I think that the code sample should show the code that suits best for the simplest use case.