nest documentation for how to find device_id
The device_id is *very* hidden by the GUI and hard to find without digging in to code.  One of the examples in this doc shows it being used, but never how to find it. My use case was Node-RED automation.

By the way - if any devs have better ways of finding this information, please decline this commit and share the better way. Just writing from a noob's perspective who figured it out the hard way.