Update 2021-04-07-release-20214.markdown to highlight changes to zwave_js_event
Currently, the Z-Wave JS Update section of the release notes makes no mention of the fact that the zwave_js_event event has been split and completely reworked.  While the information is included in the "Breaking Changes" section, I think the scope and impact of this change justifies highlighting it in the main section that people are more likely to read.

Ideally I think the z-wave js integration documentation that I've added a lint to should also be updated with a section detailing the changes that need to be made to switch from `zwave_js_event` to `zwave_js_value_notification` for Central Scene triggers (and maybe also some other common conversions).