Update mqtt.markdown
When I followed the instructions it made it seem like I can configure host/port with GUI and then add the TLS options via configuration file and that did not work. When there is any GUI configuration the mqtt configuration in the yaml file is ignored. This is not obvious, hopefully the proposed addition to help documentation will be useful.