Update html5.markdown
Clarified what mail was needed in vapid_email as it recalls "see configuring the platform" and there, at point 9, was suggested to write down the mail you can find in service account tab, Firebase service account field it was implicit that the mail was that. But it's not! You have to insert your google account mail that's associated with the firebase account. Hope that helps people to not struggle like me trying ot understand why it's not working...