Putting the easiest method at the top
by far the easiest method to retrieve the access token is to use the get mi home devices app so to avoid people getting confused and use the harder method I have moved the easier method at the top. Personally, I think if there was an accordion to hide the alternative ways to retrieve the token it will make the page much cleaner as there are lots of methods but the user only needs 1 way to retrieve the token.