Update modes.markdown
When trying to set up an process of allowing the lights to decrease while the bottom was pressed I interrupted the  single mode incorrectly and thoughts triggers for the Automation could still alter the behavior of the running instance of the automation if in single. This suggestion is to help any other user that might think the same.