Delete invalid log_levels?
I tried to use 'warn' in a script but I got the error:
```Error executing script. Invalid data for call_service at pos 1: value must be one of ['critical', 'debug', 'error', 'info', 'warning'] for dictionary value @ data['level']```

I don't know if this is a documentation error with this page or if the other log_levels are just not valid in scripts.